% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{alignTags}
\alias{alignTags}
\title{Generate cell x tag count matrix from read table}
\usage{
alignTags(
  read_table,
  tag.ref,
  filter.cells = NULL,
  string.dist.method = "hamming",
  max.dist = 1
)
}
\arguments{
\item{read_table}{Data.frame containing variables "Cell", "UMI", and "Sample" as output by readTags() function}

\item{tag.ref}{Character vector of sample tag sequences used in experiment}

\item{filter.cells}{Optional: a character vector of cell barcodes to filter the read table by. Redundant if already filtered with readTags() (default: NULL)}

\item{string.dist.method}{Specify method for calculating string distance between reference tags and tag reads when correcting for sequencing errors. See ?stringdist for options (default: hamming)}

\item{max.dist}{Specify maximum string distance allowed for correcting sequencing errors in sample tags (default: 1)}
}
\value{
Sparse matrix of cell x sample tag counts
}
\description{
Tally total UMI counts of each sample barcode per cell in read table to produce a tag count matrix.
Updated version of MULTIseq.align() from original deMULTIplex package.
}
\examples{
> head(read_table, n = 3)
            Cell      UMI   Sample
CTTAGGCCTGAACATA TTTCACAC TGAGACCT
TCTGAGCCTAAACTGA CAAAGAGG CCACAATG
TTCTAGACTGAATTGA GATACGCA TGAGACCT

tag_mtx <- alignTags(read_table,
                     tag.ref,
                     filter.cells = exp2_cells)

}
