% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiseq_oligos}
\alias{multiseq_oligos}
\title{MULTI-seq oligo sequence}
\format{
An object of class \code{"character"}
}
\usage{
data(multiseq_oligos)
}
\description{
Current version of MULTI-seq oligos. Used as reference for aligning tag fastq files.
}
\examples{
data(multiseq_oligos)
}
\references{
McGinnis, Patterson, et al., Nature methods 16.7 (2019): 619-626.
}
\keyword{datasets}
