% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{readTags}
\alias{readTags}
\title{Preprocess Tag Reads}
\usage{
readTags(
  dir,
  name,
  barcode.type = c("MULTIseq", "MULTI-ATAC", "custom"),
  assay = c("RNA", "ATAC", "Multiome"),
  filter.cells = NULL,
  fastq.A = NA,
  fastq.B = NA,
  pos.cbc,
  pos.sbc,
  pos.umi
)
}
\arguments{
\item{dir}{Path to directory containing tag FASTQ files}

\item{name}{Prefix of FASTQ files for the tag library to be processed}

\item{barcode.type}{Type of barcoding used: "MULTIseq", "MULTI-ATAC" or "custom", if using "custom", please specify fastq.A, fastq.B, pos.cbc, pos.sbc, pos.umi}

\item{assay}{10X Genomics single-cell assay used: "RNA", "ATAC", "Multiome"}

\item{filter.cells}{Optional: a character vector of cell barcodes to filter the read table by (default: NULL)}

\item{fastq.A}{Optional: provide path to FASTQ containing cell barcodes}

\item{fastq.B}{Optional: provide path to FASTQ containing sample tags}

\item{pos.cbc}{Optional: numeric vector specifying start & end indices of cell barcode sequence in FASTQ A}

\item{pos.sbc}{Optional: numeric vector specifying start & end indices of sample tag sequence in FASTQ B}

\item{pos.umi}{Optional: numeric vector specifying start & end indices of UMI sequence in FASTQ A or B (depends on barcode & assay)}
}
\value{
Data.frame containing variables "Cell", "UMI", and "Sample"
}
\description{
Loads the FASTQ files produced by sequencing MULTIseq or MULTI-ATAC barcode libraries. \cr
Read pairs are parsed to extract 10X cell barcodes, MULTI* sample tags, and UMIs, and these are assembled into a read table. \cr
Read table can be optionally filtered by a set of known cell barcodes (i.e. cells identified by Cell Ranger). \cr
Warning: This function is memory intensive for large FASTQ files. \cr
Note: This is an updated version of MULTIseq.preprocess() from original deMULTIplex package.
}
\examples{
~/Experiment2
| Exp2MULTI_S3_L002_R1_001.fastq.gz
| Exp2MULTI_S3_L002_R2_001.fastq.gz

read_table <- readTags(dir = "~/Experiment2",
                       name = "Exp2MULTI",
                       barcode.type = "MULTIseq",
                       assay = "RNA",
                       filter.cells = exp2_cells)
}
