% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{revComplement}
\alias{revComplement}
\title{Generate Reverse Complement Sequence}
\usage{
revComplement(cells, cbLength = 16, keepFormat = FALSE)
}
\arguments{
\item{cells}{A character vector of cell barcodes (may include non-barcode formatting characters)}

\item{cbLength}{Length of cell barcode (default: 16)}

\item{keepFormat}{Whether or not to keep non-barcode formatting characters (e.g. "AGTCAGTCAGTCAGTC-1" or "Sample#AGTCAGTCAGTCAGTC")}
}
\value{
A character vector
}
\description{
Takes cell barcodes formatted by Cell Ranger and/or ArchR and returns reverse complement of each cell barcode sequence.
Depending on sequencer used, this may be needed to properly align cell barcodes.
}
