% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tagCallHeatmap}
\alias{tagCallHeatmap}
\title{Classification x Tag Heatmap}
\usage{
tagCallHeatmap(
  tag_mtx,
  calls,
  log = TRUE,
  colLow = "white",
  colHigh = "#c73647"
)
}
\arguments{
\item{tag_mtx}{Tag count matrix as generated by alignTags()}

\item{calls}{Character vector of classification results}

\item{log}{logical; Whether to log-scale tag counts (default: TRUE)}

\item{colLow}{Color representing the minimum value of the heatmap}

\item{colHigh}{Color representing the maximum value of the heatmap}
}
\description{
After completing classifications, plot sample tag x classification heatmap, with values representing the average tag count for each classified sample. Helpful visualization for troubleshooting your dataset.
}
