% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tagHist}
\alias{tagHist}
\title{Tag-Count Histograms
Plot log-scaled UMI histograms for each unique tag (in clean datasets, these will appear bimodal) and total UMI counts. \cr
Helpful for visualizing tag data and diagnosing classification problems.}
\usage{
tagHist(
  tag_mtx,
  minUMI = 10,
  plotnUMI = TRUE,
  select = NULL,
  scale_y_sqrt = FALSE,
  bins = 100,
  colors = NULL
)
}
\arguments{
\item{tag_mtx}{Tag count matrix as generated by alignTags()}

\item{minUMI}{Exclude cells from plotting that do not have this minimum number of total tag UMIs (default: 10)}

\item{plotnUMI}{logical; Whether or not to plot log-scaled nUMI as well (default: TRUE)}

\item{select}{Numeric vector identifying specific barcodes (columns of tag_mtx) to plot}

\item{scale_y_sqrt}{logical; Whether or not to square-root-scale the y-axis to improve visualization (default: FALSE)}

\item{bins}{Number of bins passed on to geom_histogram() (default: 100)}

\item{colors}{Values passed on scale_fill_manual() if desired}
}
\description{
Tag-Count Histograms
Plot log-scaled UMI histograms for each unique tag (in clean datasets, these will appear bimodal) and total UMI counts. \cr
Helpful for visualizing tag data and diagnosing classification problems.
}
