% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stoeckius_pbmc}
\alias{stoeckius_pbmc}
\title{Tag count matrix of 8-donor dataset from Stoeckius et al.}
\format{
An object of class \code{"matrix"}
}
\usage{
data(stoeckius_pbmc)
}
\description{
Tag count matrix of PBMC from 8 donors from Stoeckius et al., 2018
}
\examples{
data(stoeckius_pbmc)
}
\references{
Stoeckius et al., Genome biology 19.1 (2018): 1-12.
}
\keyword{datasets}
